chore(connect): reorganize types imports
