+++ packages/connect/src/api/bitcoin/signtxLegacy.ts
@@ -1,7 +1,7 @@
 // origin: https://github.com/trezor/connect/blob/develop/src/js/core/methods/helpers/signtx-legacy.js
 
 import { PROTO, ERRORS } from '../../constants';
-import type { SignedTransaction } from '../../types/api/signTransaction';
+import type { SignedTransaction } from '../../types/api/bitcoin';
 import type { SignTxHelperParams, SignTxHelperProps } from './signtx';
 
 const requestPrevTxInfo = ({
